show correct error messages when deleting creations
