+++ app/web-frontend/src/api/errorMap.js
@@ -1,5 +1,7 @@
 export default Object.freeze({
   'not found': 'Oops! something went wrong',
+  // for creation
+  'creation has ongoing litigation process': 'Cannot delete! Creation has an ongoing litigation process',
   // for litigation
   '"litigation_end" must be greater than "ref:litigation_start"': 'End date must be greater than start date',
   'creation already assigned to a litigation': 'A litigation for this creation already exists',
