+++ app/web-frontend/src/api/requests.js
@@ -15,7 +15,10 @@ const request = async (url = '', data = {}, method = '') => {
     ...(['POST', 'PATCH'].includes(method) && {
       body: JSON.stringify(data),
     }),
-  }).then((x) => (method !== 'DELETE' ? x?.json() : x));
+  }).then((x) => {
+    if (method === 'DELETE' && x.ok) return x;
+    return x?.json();
+  });
 
   if (response?.code >= 400) {
     throw new Error(errorMap[response?.message?.toLowerCase()] || response?.message);
